fixed errors

+++ src/main/java/com/demo/TeamPlayerMicroService/DTO/PlayerStatsDTO.java
@@ -4,10 +4,10 @@ import lombok.Data;
 
 @Data
 public class PlayerStatsDTO {
-    private Long playerId;
-    private Integer runs;
-    private Integer wickets;
-    private Integer catches;
-    private Integer ballsFacedByBatsman;
-    private Integer runsConcededByBowler;
+    private int playerId;
+    private int runs;
+    private int wickets;
+    private int catches;
+    private int ballsFacedByBatsman;
+    private int runsConcededByBowler;
 }
