+++ src/main/java/com/demo/TeamPlayerMicroService/DTO/TeamPlayerDTO.java
@@ -4,12 +4,12 @@ import lombok.Data;
 
 @Data
 public class TeamPlayerDTO {
-    private Long id;
-    private Long playerId;
+    private int id;
+    private int playerId;
     private String playerName;
-    private Integer playerRuns;
-    private Integer playerWickets;
-    private Integer playerCatches;
-    private Integer ballsFacedByBatsman;
-    private Integer runsConcededByBowler;
+    private int playerRuns;
+    private int playerWickets;
+    private int playerCatches;
+    private int ballsFacedByBatsman;
+    private int runsConcededByBowler;
 }
